+++ README.md
@@ -1,22 +1,72 @@
-# next-template
+# Smart Contract Vulnerability Classification (SCVC)
 
-A Next.js 13 template for building apps with Radix UI and Tailwind CSS.
+Welcome to the **Smart Contract Vulnerability Classification (SCVC)** project! This repository provides a comprehensive framework designed to identify, categorize, and assess vulnerabilities in smart contracts. Built on the foundation of BLOCKDEF, SCVC aims to enhance the security and reliability of EVM compatible Smart Contracts.
 
-## Usage
+## 🚀 Overview
 
-```bash
-npx create-next-app -e https://github.com/shadcn/next-template
-```
+**SCVC** is an advanced tool for smart contract vulnerability analysis. It categorizes vulnerabilities, assesses their impact, and provides detailed information to help developers and auditors ensure their contracts are secure. Whether you're developing a new smart contract or auditing an existing one, SCVC is your go-to solution for robust vulnerability management.
 
-## Features
+## 📦 Features
 
-- Next.js 13 App Directory
-- Radix UI Primitives
-- Tailwind CSS
-- Icons from [Lucide](https://lucide.dev)
-- Dark mode with `next-themes`
-- Tailwind CSS class sorting, merging and linting.
+- **Comprehensive Classification**: Identify and classify various types of vulnerabilities, setting new standards in vulnerability categorization.
+- **Detailed Assessment**: Evaluate the potential impact and risks associated with each vulnerability, pioneering new methodologies for risk assessment.
+- **Integration Ready**: Easily integrate with your existing development and audit workflows.
+- **User-Friendly Interface**: Intuitive and easy-to-use framework for efficient analysis.
 
-## License
 
-Licensed under the [MIT license](https://github.com/shadcn/ui/blob/main/LICENSE.md).
+## 🛠️ Getting Started
+
+To get started with SCVC, follow these steps:
+
+### Prerequisites
+
+- **Node.js**: Ensure you have Node.js installed. [Download Node.js](https://nodejs.org/)
+- **Git**: Make sure Git is installed on your machine. [Download Git](https://git-scm.com/)
+
+### Installation
+
+1. **Clone the Repository**:
+
+    ```bash
+    git clone https://github.com/blockdef/scvc.git
+    cd scvc
+    ```
+
+2. **Install Dependencies**:
+
+    ```bash
+    npm install
+    ```
+
+3. **Run the Application**:
+
+    ```bash
+    npm start
+    ```
+
+## 🤝 Contributing
+
+We welcome contributions to SCVC! If you'd like to contribute, please follow these steps:
+
+1. **Fork the Repository**: Create a personal copy of the repository.
+2. **Create a Branch**: Make your changes in a new branch.
+3. **Submit a Pull Request**: Describe your changes and submit a pull request.
+
+Please review our [contribution guidelines](CONTRIBUTING.md) for more details on how to contribute.
+
+## 📝 License
+
+SCVC is licensed under the [MIT License](LICENSE). See the [LICENSE](LICENSE) file for details.
+
+## 📄 Documentation
+
+For detailed documentation, visit the [documentation page](https://blockdef.mintlify.app/introduction).
+
+## 🛠️ Support
+
+If you encounter any issues or have questions, please open an issue on GitHub or contact our support team at [support@scvc.org](mailto:support@blockdef.com).
+
+
+---
+
+Built with ❤️ by the SCVC team and contributors. Explore more at [BLOCKDEF](https://blockdef.dev).
