+++ content/header-buttons.ts
@@ -1,0 +1,19 @@
+import { ButtonConfig } from "@/types/button"
+
+export const headerButtonsData: ButtonConfig[] = [
+  {
+    variant: 'outline',
+    className: 'dark:bg-red w-full font-bold bg-accent border-transparent hover:bg-[#3333] md:w-auto mx-[0.1rem] -ml-[7.5rem] dark:hover:bg-white hover:text-black',
+    children: 'About'
+  },
+  {
+    variant: 'outline',
+    className: 'w-full font-bold bg-accent border-transparent hover:bg-[#3333] md:w-auto mx-[0.1rem] dark:hover:bg-white hover:text-black',
+    children: 'Support'
+  },
+  {
+    variant: 'outline',
+    className: 'w-full font-bold bg-accent border-transparent hover:bg-[#3333] md:w-auto mx-[0.1rem] dark:hover:bg-white hover:text-black',
+    children: 'Changelog'
+  }
+];
